Add more test cases to entity parser utility
