+++ src/test/resources/metafile-no-non-scannables.json
@@ -1,0 +1,38 @@
+{
+  "po_box": "XXXX",
+  "jurisdiction": "SSCS",
+  "delivery_date": "23-02-2017 00:00:00.100000",
+  "opening_date": "24-02-2017 00:00:00.010000",
+  "zip_file_createddate": "24-02-2017 00:00:00.001000",
+  "zip_file_name": "1_24-02-2017-00-00-00.zip",
+  "scannable_items": [
+    {
+      "document_control_number": "1111001",
+      "scanning_date": "24-02-2017 00:00:00.000100",
+      "ocr_accuracy": "string",
+      "manual_intervention": "string",
+      "next_action": "return",
+      "next_action_date": "24-02-2017 00:00:00.000010",
+      "ocr_data": "R0lGODlhAQABAIAAAAAAAAAAACH5BAAAAAAALAAAAAABAAEAAAICTAEAOw==",
+      "file_name": "1111001.pdf",
+      "notes": "The document had ink spilled on it"
+    },
+    {
+      "document_control_number": "1111002",
+      "scanning_date": "24-02-2017 00:00:00.000001",
+      "manual_intervention": "string",
+      "next_action": "forward",
+      "next_action_date": "25-02-2017 00:00:00.000000",
+      "file_name": "1111002.pdf",
+      "notes": "Cheque will be forwarded to court to bank it"
+    }
+  ],
+  "payments": [
+    {
+      "document_control_number": "1111002",
+      "method": "Cheque",
+      "amount": "100.00",
+      "currency": "GBP"
+    }
+  ]
+}
